Test: Remove entry and output from webpack config
When these are present in the webpack config that is passed to karma the build gets much bigger and node is struggling. Since they'll never be needed when testing (every single test file is an entry point) they can safely be deleted from the webpack config after building it from the javascript-env.js file.